Date: 20150809
File: portage-latest.tar.xz
Verify result:
gpg: Signature made Sun Aug  9 02:55:29 2015 CEST using RSA key ID C9189250
gpg: Good signature from "Gentoo Portage Snapshot Signing Key (Automated Signing Key)" [ultimate]

gpgsig:
-----BEGIN PGP SIGNATURE-----
Version: GnuPG v2

iQIcBAABAgAGBQJVxqUBAAoJEOxZDurJGJJQ7zIP/0mQlCyHDvWHOPAHOTm4D5op
ESflh9E46voAD4lZ3Z9Q9/aukUw6Rlu8LYjyZKJfgynQ0A/vX1iv/f1RmmNfS72S
PfT6CavWLuYjHl9ml/YTm8EHvvKIJfWtlOrOGVNatZvgsuUD5qE0dHjoZ6/SN1kq
pbMCLrc2Z7/8YuuEJjQpwz437vfBS9IUlTpLnNDGLY+z+N2uPdTsuAvKKxxmW4ZV
7YQXYD32JSBWCuBJvYQivF6oMg4Rp+bIvru9ie4+w6fblTSfcn85TmXGF+mLXZ3l
Y567EueFz7JO3WuX/1Pw7L/XStq1fQ+6bnd7qqHkbDoO3Y+BdBVp2hVKXDD9EjMn
hruyYWBx27L1vpBrKMp3GN9jKWxHBghMYPXcmTyRvpsso/x1Vs2wekkuViq1g3jv
+XzOlgXA+t2Nu0m3j4cIienaPmqaBCC2cwRQaiwbp1SJ9x2Huwd8x+CMHt8rVzpH
Bp/tkYVq99MCkwQwxBfra253IQ37F0kbZ9h6V/uGIcQGytKZwmkBuWEEW61847p/
k1vYdwOBlDk6h+V3YR3Q+jW/ZW9ci5mvwTSgJARslGL4lbP6YcWUB+n0iEdLaNYI
EQly5wI9y1rnegWoq1NAuwE51sm38uULORbLYF6pg7RkSVc5Ir+IYhb8UeLEhIr9
gtBT8Fe6z7Qka13Umkq1
=DCDu
-----END PGP SIGNATURE-----

+++ dev-java/javacup/javacup-0.11b_beta20150326.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2015 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/dev-java/javacup/javacup-0.11b_beta20150326.ebuild,v 1.1 2015/08/07 19:58:56 sping Exp $
+# $Header: /var/cvsroot/gentoo-x86/dev-java/javacup/javacup-0.11b_beta20150326.ebuild,v 1.4 2015/08/08 13:07:19 sping Exp $
 
 EAPI="5"
 
@@ -22,36 +22,41 @@ SRC_URI="http://www2.cs.tum.edu/projects/cup/releases/java-cup-src-${MY_PV}.tar.
 LICENSE="GPL-2"
 SLOT="0"
 KEYWORDS="~amd64 ~ppc ~ppc64 ~x86 ~amd64-fbsd ~x86-fbsd ~x64-freebsd ~x86-freebsd ~amd64-linux ~x86-linux ~ppc-macos ~x64-macos ~x86-macos ~sparc-solaris ~sparc64-solaris ~x64-solaris ~x86-solaris"
-IUSE="userland_BSD"
+IUSE="userland_BSD system-jflex"
 
 # find for bug #214664
 DEPEND=">=virtual/jdk-1.5
-	dev-java/jflex:0
-	!userland_BSD? ( >=sys-apps/findutils-4.4 )"
-RDEPEND=">=virtual/jre-1.5
-		>=dev-java/ant-core-1.7.0"
+	system-jflex? ( dev-java/jflex:0 )
+	!userland_BSD? ( >=sys-apps/findutils-4.4 )
+	>=dev-java/ant-core-1.7.0:0"
+RDEPEND=">=virtual/jre-1.5"
 
 S="${WORKDIR}"
 
 src_prepare() {
 	epatch "${FILESDIR}"/${P}-build-xml.patch
+	epatch "${FILESDIR}"/${P}-build-xml-svn.patch
 
 	find . -name '*.class' -delete || die
 
-	rm bin/JFlex.jar || die
-	java-pkg_jar-from --build-only jflex JFlex.jar bin/JFlex.jar
+	if use system-jflex; then  # break the circular dependency
+		rm bin/JFlex.jar || die
+		java-pkg_jar-from --build-only jflex JFlex.jar bin/JFlex.jar
+	fi
 
 	java-ant_rewrite-classpath
 }
 
 src_compile() {
-	eant -Dgentoo.classpath="$(java-pkg_getjars ant-core)"
+	local classpath="$(java-pkg_getjars --build-only ant-core)"
+
+	eant -Dgentoo.classpath="${classpath}"
 	rm bin/java-cup-11.jar || die
 	cp dist/java-cup-11b.jar bin/java-cup-11.jar || die
 	eant clean
 
 	einfo "Recompiling with newly generated javacup"
-	eant -Dgentoo.classpath="$(java-pkg_getjars ant-core)"
+	eant -Dgentoo.classpath="${classpath}"
 	use doc && javadoc -sourcepath src/ java_cup -d javadoc
 }
